change char's xadvance property from uint16 to float
(cherry picked from commit ebd50a106b43b3f490e60fc74966f7eebfa155fe)
(cherry picked from commit ef5452ff7c2e601f5c27185c27f3b74ba4880be3)

+++ Source/LGUI/Public/Core/LGUIFontData.h
@@ -42,7 +42,7 @@ public:
 	uint16 height = 0;
 	int16 xoffset = 0;
 	int16 yoffset = 0;
-	uint16 xadvance = 0;
+	float xadvance = 0;
 	float uv0X = 0;
 	float uv0Y = 0;
 	float uv3X = 0;
